+++ CHANGELOG.md
@@ -8,6 +8,20 @@
 
 -   Change hwp_angle calculation to the complex domain [#433](https://github.com/litebird/litebird_sim/pull/433)
 
+# Version 0.16.0
+
+-   Add pre-commit hook to keep syntax compliant with `py3.10+` [#461](https://github.com/litebird/litebird_sim/pull/461)
+
+-   Fix issue [#459](https://github.com/litebird/litebird_sim/issues/459)
+
+-   Fix issue [#457](https://github.com/litebird/litebird_sim/issues/457)
+
+-   Fix incorrect handling of HWP with precomputed pointing [#456](https://github.com/litebird/litebird_sim/pull/456)
+
+-   Improve the documentation [#450](https://github.com/litebird/litebird_sim/pull/450)
+
+-   Import Yusuke Takase's grasp2alm code [#447](https://github.com/litebird/litebird_sim/pull/447)
+
 # Version 0.15.3
 
 -   Fix the computation of HWP angles [#444](https://github.com/litebird/litebird_sim/pull/444)